backend/requirements.txt

fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
geoalchemy2==0.14.2
shapely==2.0.2
pydantic==2.5.3
pydantic-settings==2.1.0
alembic==1.13.1
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
httpx==0.26.0
pgvector==0.2.4
python-jose[cryptography]==3.3.0
numpy<2.0.0
pillow==11.2.1
email-validator==2.1.0
python-dotenv==1.0.1
